+++ pyppeteer/navigator_watcher.py
@@ -5,17 +5,12 @@
 
 import asyncio
 import concurrent.futures
-
 from typing import Any, Awaitable, Dict, List, Union
-from typing import TYPE_CHECKING
 
 from pyppeteer import helper
 from pyppeteer.errors import TimeoutError
 from pyppeteer.frame_manager import FrameManager, Frame
 from pyppeteer.util import merge_dict
-
-if TYPE_CHECKING:
-    from typing import Set  # noqa: F401
 
 
 class NavigatorWatcher:
@@ -72,13 +67,18 @@ class NavigatorWatcher:
         if options.get('waitUntil') == 'documentloaded':
             import logging
             logging.getLogger(__name__).warning(
-                '`documentloaded` option is no logner supported. '
+                '`documentloaded` option is no longer supported. '
                 'Use `domcontentloaded` instead.')
         _waitUntil = options.get('waitUntil', 'load')
         if isinstance(_waitUntil, list):
             waitUntil = _waitUntil
         elif isinstance(_waitUntil, str):
             waitUntil = [_waitUntil]
+        else:
+            raise TypeError(
+                '`waitUntil` option should be str or list of str, '
+                f'but got type {type(_waitUntil)}'
+            )
         self._expectedLifecycle: List[str] = []
         for value in waitUntil:
             protocolEvent = pyppeteerToProtocolLifecycle.get(value)