+++ pyppeteer/util.py
@@ -7,12 +7,12 @@ import gc
 import socket
 from typing import Dict, Optional
 
-from pyppeteer.chromium_downloader import check_chromium, chromium_excutable
+from pyppeteer.chromium_downloader import check_chromium, chromium_executable
 from pyppeteer.chromium_downloader import download_chromium
 
 __all__ = [
     'check_chromium',
-    'chromium_excutable',
+    'chromium_executable',
     'download_chromium',
     'get_free_port',
 ]